Pegasus install Darwin

+++ darwin/functions/overrides/helperFunctions.sh
@@ -1,6 +1,5 @@
 #!/bin/bash
 
-
 safeDownload() {
 	local name="$1"
 	local url="$2"
@@ -15,26 +14,33 @@ safeDownload() {
 		echo "- $showProgress"
 		echo "- $headers"
 	fi
-	
+
 	request=$(curl -w $'\1'"%{response_code}" --fail -L "$url" -H "$headers" -o "$outFile.temp" 2>&1 && echo $'\2'0 || echo $'\2'$?)
 
 	returnCodes="${request#*$'\1'}"
 	httpCode="${returnCodes%$'\2'*}"
 	exitCode="${returnCodes#*$'\2'}"
-	
+
 	if [ "$httpCode" = "200" ] && [ "$exitCode" == "0" ]; then
 		#echo "$name downloaded successfully";
 		mv -v "$outFile.temp" "$outFile" &>/dev/null
 		volumeName=$(yes | hdiutil attach "$outFile" | grep -o '/Volumes/.*$')
 
-		cp -r "$volumeName"/*.app "$HOME/Applications/EmuDeck/"		
+		if [ -z "$volumeName" ]; then
+			unzip "$outFile";
+			volumeName="$HOME/Applications/EmuDeck/$outFile"
+		fi
 
-		appName=$(find "$volumeName" -name "*.app" -exec basename {} \;)				
-		chmod +x  "$HOME/Applications/EmuDeck/$appName"	
-		
-		find "$HOME/Applications/EmuDeck/" -maxdepth 1 -name "*.app" -exec ln -s {} /Applications/ \;		
+		cp -r "$volumeName"/*.app "$HOME/Applications/EmuDeck/"
+
+		appName=$(find "$volumeName" -name "*.app" -exec basename {} \;)
+		chmod +x  "$HOME/Applications/EmuDeck/$appName"
+
+		find "$HOME/Applications/EmuDeck/" -maxdepth 1 -name "*.app" -exec ln -s {} /Applications/ \;
 		#chmod +x  "/Applications/$appName"
-		hdiutil detach "$volumeName" && rm -rf "$outFile"
+		if [ -n "$volumeName" ]; then
+			hdiutil detach "$volumeName" && rm -rf "$outFile"
+		fi
 		return 0
 	else
 		#echo "$name download failed"
